+++ Sources/BugsnagPerformance/Public/BugsnagPerformanceConfiguration.mm
@@ -53,6 +53,7 @@ static NSString *defaultEndpoint = @"https://otlp.bugsnag.com/v1/traces";
     auto releaseStage = getSharedConfigValue(@"releaseStage");
     auto enabledReleaseStages = getSharedConfigArray(@"enabledReleaseStages");
     
+    auto serviceName = BSGDynamicCast<NSString>(bugsnagPerformanceConfiguration[@"service.name"]);
     auto endpoint = BSGDynamicCast<NSString>(bugsnagPerformanceConfiguration[@"endpoint"]);
     auto autoInstrumentAppStarts = BSGDynamicCast<NSNumber>(bugsnagPerformanceConfiguration[@"autoInstrumentAppStarts"]);
     auto autoInstrumentViewControllers = BSGDynamicCast<NSNumber>(bugsnagPerformanceConfiguration[@"autoInstrumentViewControllers"]);
@@ -68,6 +69,9 @@ static NSString *defaultEndpoint = @"https://otlp.bugsnag.com/v1/traces";
         configuration.releaseStage = releaseStage;
     }
     configuration.enabledReleaseStages = [NSSet setWithArray: enabledReleaseStages ?: @[]];
+    if (serviceName) {
+        configuration.serviceName = serviceName;
+    }
     if (endpoint) {
         configuration.endpoint = [[NSURL alloc] initWithString: endpoint];
     }